backend/src/requirements.txt
aioredis==1.3.1
asgiref==3.2.3
astroid==2.3.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
coverage==5.1
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.3
django-allauth==0.41.0
django-cors-headers==3.2.1
django-debug-toolbar==2.2
django-google-maps==0.12.1
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.4.0
msgpack==0.6.2
numpy==1.18.3
oauthlib==3.1.0
packaging==20.4
phonenumbers==8.12.1
Pillow==7.0.0
pluggy==0.13.1
postgis==1.0.4
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
py==1.8.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==5.4.3
pytest-cov==2.10.0
pytest-django==3.9.0
pytest-pythonpath==0.7.3
python-memcached==1.59
python-resize-image==1.1.19
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rest-framework-cache==0.1
returns==0.14.0
service-identity==18.1.0
six==1.14.0
sorl-thumbnail==12.6.3
sorl-thumbnail-serializer-field==0.2.1
SQLAlchemy==1.3.17
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.8
wcwidth==0.2.4
wrapt==1.11.2
zope.interface==5.1.0
